third_party/blink/web_tests/fast/peerconnection/RTCRtpSender-setParameters-expected.txt

This is a testharness.js-based test.
[FAIL] setParameters() with large maxBitrate changes getParameters() returned values
  assert_equals: expected 4294967295 but got 2147483647
[FAIL] setParameters() to a codec with munging
  promise_test: Unhandled rejection with value: object "InvalidModificationError: Failed to execute 'setParameters' on 'RTCRtpSender': Attempted to use an unsupported codec for layer 0"
Harness: the test ran to completion.
